Refactoring: Move ISharedPreferenceService to Model layer. #56
