+++ app-model/src/main/java/com/paper/model/ISharedPreferenceService.kt
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.paper.domain
+package com.paper.model
 
 interface ISharedPreferenceService {
